Refatorando deixando com menos classes
