--- app/src/main/java/br/com/caelum/casadocodigo/repository/AutorRepository.kt
@@ -1,8 +1,0 @@
-package br.com.caelum.casadocodigo.repository
-
-import br.com.caelum.casadocodigo.model.Autor
-import br.com.caelum.casadocodigo.web.AutorWebClient
-
-class AutorRepository(private val client: AutorWebClient) {
-    fun salva(autor: Autor) = client.salva(autor)
-}
