+++ app/src/main/java/br/com/caelum/casadocodigo/viewmodel/AutorViewModel.kt
@@ -1,15 +1,19 @@
 package br.com.caelum.casadocodigo.viewmodel
 
+import android.util.Log
 import android.util.Patterns
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import br.com.caelum.casadocodigo.model.Autor
-import br.com.caelum.casadocodigo.repository.AutorRepository
-import br.com.caelum.casadocodigo.web.AutorWebClient
-import br.com.caelum.casadocodigo.web.InicializadorDeRetrofit
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import retrofit2.Retrofit
+import retrofit2.converter.jackson.JacksonConverterFactory
+import retrofit2.http.Body
+import retrofit2.http.POST
 
-class AutorViewModel private constructor(private val autorRepository: AutorRepository) : ViewModel() {
+class AutorViewModel : ViewModel() {
     private val nomeValido = MutableLiveData<String>()
     private val githubLinkValido = MutableLiveData<String>()
     val autorValido = MutableLiveData<Autor>()
@@ -48,12 +52,33 @@ class AutorViewModel private constructor(private val autorRepository: AutorRepos
         }
     }
 
-    fun salva(autor: Autor) = autorRepository.salva(autor)
+    fun salva(autor: Autor) {
+        val retrofit = Retrofit.Builder()
+            .baseUrl("https://8b27b7a2.ngrok.io")
+            .addConverterFactory(JacksonConverterFactory.create())
+            .build()
 
-    object Factory : ViewModelProvider.Factory {
-        private val autorWebClient = AutorWebClient(InicializadorDeRetrofit.retrofit)
-        private val autorRepository = AutorRepository(autorWebClient)
+        val autorService = retrofit.create(AutorService::class.java)
 
-        override fun <T : ViewModel?> create(modelClass: Class<T>): T = AutorViewModel(autorRepository) as T
+        autorService.salva(autor).enqueue(object : Callback<Unit?> {
+            override fun onFailure(call: Call<Unit?>, t: Throwable) {
+                Log.e("ERRO", "Não comunicou com o servidor", t)
+            }
+
+            override fun onResponse(call: Call<Unit?>, response: Response<Unit?>) {
+                if (response.isSuccessful) {
+                    Log.i("SUCESSO", "autor ${autor.nome} salvo")
+                } else {
+                    Log.i("FALHA", "autor ${autor.nome} não rolou ${response.raw()}")
+                }
+            }
+
+        })
     }
+
+    private interface AutorService {
+        @POST("/api/autor")
+        fun salva(@Body autor: Autor): Call<Unit?>
+    }
+
 }
