--- app/src/main/java/br/com/caelum/casadocodigo/web/InicializadorDeRetrofit.kt
@@ -1,11 +1,0 @@
-package br.com.caelum.casadocodigo.web
-
-import retrofit2.Retrofit
-import retrofit2.converter.jackson.JacksonConverterFactory
-
-object InicializadorDeRetrofit {
-    val retrofit = Retrofit.Builder()
-        .baseUrl("https://439eb8e8.ngrok.io")
-        .addConverterFactory(JacksonConverterFactory.create())
-        .build()
-}
